feat: 타입 변경
int 타입으로 파라미터를 받아도 double 타입으로 계산되도록 수정

+++ src/lv3/App.java
@@ -5,13 +5,13 @@ import java.util.Scanner;
 public class App {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
-        ArithmeticCalculator<Integer> calculator = new ArithmeticCalculator<>();
+        ArithmeticCalculator<Double> calculator = new ArithmeticCalculator<>();
 
         System.out.print("첫 번째 정수를 입력하세요: ");
-        int firstNumber = sc.nextInt();
+        double firstNumber = sc.nextInt();
 
         System.out.print("두 번째 정수를 입력하세요: ");
-        int secondNumber = sc.nextInt();
+        double secondNumber = sc.nextInt();
 
         System.out.print("사칙연산 기호를 입력하세요: ");
         char operator = sc.next().charAt(0);
